Proper following of user
the JSON returned by the search may include retweets by tweeters other than the original tweeter. Retweeting these users is fine, since twitter automatically returns the original tweet, however the function to follow a user  originally would follow the retweeter, making the program user ineligible for the giveaway. The fixed code tries to get an entry from retweeted_status that only shows up in retweeted tweets, and contains information about the original tweeter. If this fails (the tweet is not a retweet), it continues as normal.